+++ src/navigation/BirthdayMessagesNavigator.tsx
@@ -1,22 +1,22 @@
-import { BirthdaysStackParamList } from "@/src/navigation/NavigationTypes";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import React from "react";
-import BirthdayMessages from "../screens/birthdaymessages/BirtdayMessages";
-import CreateMessage from "../screens/birthdaymessages/CreateMessage";
-import HandleMessage from "../screens/birthdaymessages/HandleMessage";
+// import { BirthdaysStackParamList } from "@/src/navigation/NavigationTypes";
+// import { createNativeStackNavigator } from "@react-navigation/native-stack";
+// import React from "react";
+// import BirthdayMessages from "../screens/birthdaymessages/BirtdayMessages";
+// import CreateMessage from "../screens/home/CreateMessage";
+// import HandleMessage from "../screens/birthdaymessages/HandleMessage";
 
-const Stack = createNativeStackNavigator<BirthdaysStackParamList>();
+// const Stack = createNativeStackNavigator<BirthdaysStackParamList>();
 
-export default function BirthdaysMessagesStackNavigator() {
-  return (
-    <Stack.Navigator>
-      <Stack.Screen
-        name="BirthdayMessages"
-        component={BirthdayMessages}
-        options={{ headerShown: false }}
-      />
-      <Stack.Screen name="CreateMessage" component={CreateMessage} />
-      <Stack.Screen name="HandleMessage" component={HandleMessage} />
-    </Stack.Navigator>
-  );
-}
+// export default function BirthdaysMessagesStackNavigator() {
+//   return (
+//     <Stack.Navigator>
+//       <Stack.Screen
+//         name="BirthdayMessages"
+//         component={BirthdayMessages}
+//         options={{ headerShown: false }}
+//       />
+//       <Stack.Screen name="CreateMessage" component={CreateMessage} />
+//       <Stack.Screen name="HandleMessage" component={HandleMessage} />
+//     </Stack.Navigator>
+//   );
+// }
