+++ src/navigation/NavigationTypes.tsx
@@ -1,6 +1,7 @@
 import { BottomTabScreenProps } from "@react-navigation/bottom-tabs";
 import { CompositeScreenProps } from "@react-navigation/native";
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
+import { HomeStack } from "./HomeNavigator";
 
 export type RootTabsParamList = {
   HomeTab: undefined;
@@ -15,8 +16,8 @@ export type ContactsStackParamList = {
   AddEditContactStack: undefined;
 };
 
-export type BirthdaysStackParamList = {
-  BirthdayMessages: undefined;
+export type HomeStackParamList = {
+  Home: undefined;
   CreateMessage: { id: string };
   HandleMessage: undefined;
 };
@@ -38,9 +39,15 @@ export type ContactsScreenProps<T extends keyof ContactsStackParamList> =
     BottomTabScreenProps<RootTabsParamList>
   >;
 
-export type BirthdaysScreenProps<T extends keyof BirthdaysStackParamList> =
+// export type BirthdaysScreenProps<T extends keyof BirthdaysStackParamList> =
+//   CompositeScreenProps<
+//     NativeStackScreenProps<BirthdaysStackParamList, T>,
+//     BottomTabScreenProps<RootTabsParamList>
+//   >;
+
+export type HomeScreenProps<T extends keyof HomeStackParamList> =
   CompositeScreenProps<
-    NativeStackScreenProps<BirthdaysStackParamList, T>,
+    NativeStackScreenProps<HomeStack, T>,
     BottomTabScreenProps<RootTabsParamList>
   >;
 
@@ -53,7 +60,7 @@ export type TestScreenProps<T extends keyof TestStackParamList> =
 export type RootScreenProps<T extends keyof RootTabsParamList> =
   CompositeScreenProps<
     BottomTabScreenProps<RootTabsParamList, T>,
-    NativeStackScreenProps<BirthdaysStackParamList>
+    NativeStackScreenProps<HomeStackParamList>
   >;
 
 // // export type RootStackScreenProps<T extends keyof HomeTabScreenProps> =
