+++ src/navigation/TestNavigator.tsx
@@ -3,7 +3,6 @@ import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import React from "react";
 import CredUser from "../screens/test-screens/CredUser";
 import Main from "../screens/test-screens/Main";
-// import Test from "../screens/test-screens/Test";
 
 const Stack = createNativeStackNavigator<TestStackParamList>();
 
@@ -15,7 +14,6 @@ export default function TestStackNavigator() {
         component={Main}
         options={{ headerShown: false }}
       />
-      {/* <Stack.Screen name="Test" component={Test} /> */}
       <Stack.Screen name="CredUser" component={CredUser} />
     </Stack.Navigator>
   );
